+++ src/store.js
@@ -1,10 +1,13 @@
 import {combineReducers, createStore} from "redux";
 import acciones from './reducers/accciones'
 
+const initialState = {
+}
+
 const reducer = combineReducers({
   acciones,
 })
 
-const store = createStore(reducer)
+const store = createStore(reducer, initialState)
 
 export default store
